DEBUG: use RTTI to determine type of C++ class instance
This commit is intended entirely for debugging purposes only. It is an
example how to use RTTI (and how to get it working) so that one can
inspect instances for their type.

Unfortunately, due to C++' name mangling, the output is not quite as easy
to read as Java's .getClass().getName()... ;-)

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>